Moved primitive UI components.
