+++ packages/interface/src/screens/settings/ExperimentalSettings.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { useNodeStore } from '../../components/device/Stores';
-import { Toggle } from '../../components/primitive';
-import { InputContainer } from '../../components/primitive/InputContainer';
+import { Toggle } from '../../../../ui/src/Toggle';
+import { InputContainer } from '../../../../ui/src/InputContainer';
 import { SettingsContainer } from '../../components/settings/SettingsContainer';
 import { SettingsHeader } from '../../components/settings/SettingsHeader';
