+++ packages/interface/src/screens/settings/GeneralSettings.tsx
@@ -1,8 +1,8 @@
 import { useBridgeQuery } from '@sd/client';
 import React from 'react';
 
-import { InputContainer } from '../../components/primitive/InputContainer';
-import Listbox from '../../components/primitive/Listbox';
+import { InputContainer } from '../../../../ui/src/InputContainer';
+import Listbox from '../../../../ui/src/Listbox';
 import { SettingsContainer } from '../../components/settings/SettingsContainer';
 import { SettingsHeader } from '../../components/settings/SettingsHeader';
