+++ packages/interface/src/screens/settings/LocationSettings.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { InputContainer } from '../../components/primitive/InputContainer';
+import { InputContainer } from '../../../../ui/src/InputContainer';
 import { SettingsContainer } from '../../components/settings/SettingsContainer';
 import { SettingsHeader } from '../../components/settings/SettingsHeader';
