+++ packages/interface/src/screens/settings/SecuritySettings.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@sd/ui';
 import React from 'react';
 
-import { InputContainer } from '../../components/primitive/InputContainer';
+import { InputContainer } from '../../../../ui/src/InputContainer';
 import { SettingsContainer } from '../../components/settings/SettingsContainer';
 import { SettingsHeader } from '../../components/settings/SettingsHeader';
